renamed master to clusterManager (#319)
