+++ src/main/kotlin/org/opensearch/replication/action/autofollow/AutoFollowClusterManagerNodeAction.kt
@@ -9,14 +9,14 @@
  * GitHub history for details.
  */
 
-package org.opensearch.replication.action.index
+package org.opensearch.replication.action.autofollow
 
 import org.opensearch.action.ActionType
 import org.opensearch.action.support.master.AcknowledgedResponse
 
-class ReplicateIndexMasterNodeAction private constructor(): ActionType<AcknowledgedResponse>(NAME, ::AcknowledgedResponse) {
+class AutoFollowClusterManagerNodeAction: ActionType<AcknowledgedResponse>(NAME, ::AcknowledgedResponse) {
     companion object {
-        const val NAME = "internal:indices/admin/plugins/replication/index/start"
-        val INSTANCE: ReplicateIndexMasterNodeAction = ReplicateIndexMasterNodeAction()
+        const val NAME = "internal:cluster:admin/plugins/replication/autofollow/update"
+        val INSTANCE = AutoFollowClusterManagerNodeAction()
     }
 }
